Does not crash at last :)
Fallback mechanism (to SampleAdNetwork logic) is working.
First bid (based on impressions worth) is OK.
Second bid (based on campaigns' priorities and first bid) is NOT OK.

Probably should insert more default cases to guide the classifier in
such cases.

Removed Publisher's popularity from the parameters for the bid - this is
irrelevant.

Lots of debug messages - so be cautious.